refactor(silent installation): ♻️ adding silent installation flag for symlinks

+++ 00-INSTALLER/02-SCRIPTS/00-symlink-files.sh
@@ -37,6 +37,9 @@ for filename in .* *; do
     elif [ "$(readlink "$targetFile")" == "$sourceFile" ]; then
       print_success "$targetFile → $sourceFile"
     else
+
+if [ -z "$SILENT_INSTALL" ]
+then
       ask_for_confirmation "'$targetFile' already exists, do you want to overwrite it?"
       if answer_is_yes; then
         rm -rf "$targetFile"
@@ -44,6 +47,11 @@ for filename in .* *; do
       else
         print_error "$targetFile → $sourceFile"
       fi
+else
+      print_info ">>> SYMLINKING SILENTLY"
+fi
+
+
     fi
   fi
 done
